Disable auto refresh
Hopefully you find this useful. Auto refresh was driving me crazy so I added a button to toggle it and track that via local storage. It defaults to off but that can easily be changed. If there is something you don't like about it let me know and I'll try and fix it up.